Fix Lux/Footcandle calculation

+++ core/modules/LightSensor.h
@@ -7,17 +7,14 @@
 
 class LightSensor
 {
-
 public:
 	LightSensor(const char * filename, uint8_t addr);
-	uint16_t getLux();
+	uint32_t getLux();
 	uint32_t getFootCandles();
 	
 private:
-
 	Adafruit_TSL2591 _tsl2591;
   	
 };
-
 
 #endif
